requirements-dev.txt
pytest==3.1.3
pytest-django==3.1.2
pytest-cov==2.5.1
flake8==3.4.1
responses==0.7.0
testfixtures==5.1.1
